Add Custom Exception Classes and embed in Controllers

See merge request spb-net-laboratory-2021/2021q3/kulakovskiie!13

+++ Module-10/WebApplicationStudents/WebApplicationStudents/Controllers/HomeworkController.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using WebApplicationStudents.Models;
 using AutoMapper;
+using WebApplicationStudents.Exceptions;
 
 namespace WebApplicationStudents.Controllers
 {
@@ -26,6 +27,8 @@ namespace WebApplicationStudents.Controllers
         [HttpGet("id")]
         public ActionResult<HomeworkBl> GetHomework(int id)
         {
+            if (id <= 0) throw new IndexLessThanZeroException($"Id {id} of Homework was less than zero");
+
             return _homeworkService.Get(id) switch
             {
                 null => NotFound(),
@@ -49,6 +52,8 @@ namespace WebApplicationStudents.Controllers
         [HttpPut("{id}")]
         public ActionResult<string> UpdateHomework(int id, Homework homework)
         {
+            if (id <= 0) throw new IndexLessThanZeroException($"Id {id} of Homework was less than zero");
+
             var homeworkBl = _mapper.Map<HomeworkBl>(homework);
             var homeworkId = _homeworkService.Edit(homeworkBl with { Id = id });
             return Ok($"api/homework/{homeworkId}");
@@ -58,6 +63,8 @@ namespace WebApplicationStudents.Controllers
         [HttpDelete("{id}")]
         public ActionResult DeleteHomework(int id)
         {
+            if (id <= 0) throw new IndexLessThanZeroException($"Id {id} of Homework was less than zero");
+
             _homeworkService.Delete(id);
             return Ok();
         }
